+++ packages/site/src/routes/[dictionaryId]/entries/print/QrCode.svelte
