+++ packages/site/src/routes/[dictionaryId]/entries/print/mock-data.ts
