+++ packages/site/src/routes/[dictionaryId]/entries/print/qrcodegen.ts
